removed gm_parser
