+++ cmake/headers.cmake
@@ -66,7 +66,7 @@ include/warthog/util/cost_table.h
 include/warthog/util/dimacs_parser.h
 include/warthog/util/experiment.h
 include/warthog/util/file_utils.h
-include/warthog/util/gm_parser.h
+include/warthog/util/gm_header.h
 include/warthog/util/helpers.h
 include/warthog/util/intrin.h
 include/warthog/util/macros.h
