+++ include/warthog/domain/labelled_gridmap.h
@@ -19,7 +19,7 @@
 #include "grid.h"
 #include <warthog/constants.h>
 #include <warthog/io/bittable_serialize.h>
-#include <warthog/util/gm_parser.h>
+#include <warthog/util/gm_header.h>
 #include <warthog/util/helpers.h>
 
 #include <cassert>
