+++ src/CMakeLists.txt
@@ -29,7 +29,6 @@ util/cost_table.cpp
 util/dimacs_parser.cpp
 util/experiment.cpp
 util/file_utils.cpp
-util/gm_parser.cpp
 util/helpers.cpp
 util/scenario_manager.cpp
 util/timer.cpp
